Ensure all opens are contained in sysroot
This is done using openat2() with RESOLVE_IN_ROOT.  To work around a
limitation of Arch Linux's musl 1.2.4-1 package, the needed definitions
from <linux/openat2.h> are open-coded.  The only opens that do not
use RESOLVE_IN_ROOT are either outside of the chroot or are before any
processes are launched.