Load extra modules necessary in 6.1.x kernel
virtio_pci_modern_dev and virtio_pci_legacy_dev are dependencies of
virtio_pci. netfs is a dependency of fscache and 9p.
Kernel config doesn't allow building those as built-in when actual
functionality is in a module, so load the extra modules manually when
necessary.

Make it optional, to still work with older kernel too.